WIP: Formalize the query checking
Currently, the WIP has Postgres checking query counts, select counts,
and update counts. All that is left on Postgres is to check the
thresholds for updates and query counts, and *maybe* the fortunes
dynamic size test impl... might hold off on that one as a todo.

+++ src/main.rs
@@ -33,7 +33,7 @@ fn main() -> VerifierResult<()> {
         database,
         concurrency_levels
             .split(',')
-            .map(|item| i32::from_str(item).unwrap())
+            .map(|item| i64::from_str(item).unwrap())
             .collect(),
     )?;
